isAdmin grabs the user's email and double checks that it's an admin in a Firebase collection
